refactor(agents): extract config composition and integrations-sync helpers
Move composeJsonConfig / decomposeJsonConfig into
json-config/agent-config-composition.ts and the syncIntegrations
orchestrator into integrations/integrations-sync.ts so this PR's
additions don't grow agents.service.ts.
